WorldTrotter v6.0 b4.10: implement a map view delegate method that should fire each time an annotation is added; set map view delegate property to self; the latter is required or the method is not called. Then decide not to use this at this time since I am adding the pins programmatically at this. Add the coordinates for each pin to an array. But the program fails with a “Class ‘MapViewController’ has no initializers” error.
Eventually, find a Stack Overflow posting (http://stackoverflow.com/questions/25811622/class-viewcontroller-has-no-initializers-in-swift) that deals with this. So it seems that some variable I added recently is not initialized. But which one? Then in Xcode, experiment with the navigator panels, and select the Issue Navigator. Aha! It shows this issue. Expanding the issue hierarchy reveals the cause as “Stored property ‘indexOfLastAnnotation’ without initial value prevents synthesized initializers”. Now I just need to figure out how to fix this.