WorldTrotter v7.0 b2.0: continue by generating a new Localizable.strings text file on the command line, but I do not see how to replace the file of the same name that is already in the WorldTrotter project. So I manually copied the new mappings to the existing base, English, and Spanish versions of this file. Then I used Google Translate to find the Spanish translation of the user location and go to pin buttons, and entered those in the Localizable.strings (Spanish) text file. Then I built the project with the Spanish scheme, and all of the labels and button titles are in Spanish!
Also (not in book exercises) I modified the bottom constraint on each of my custom buttons to offset them from the bottom layout guide. It took some experimentation and reading in the doc, but it works; now the buttons are visually separated from the tab bar.

Localize the app for German to meet the chapter 7 Bronze Challenge. Simply by following the same procedure that I did for Spanish, adding a localization in the Project settings for German, and then editing the resulting Main.strings (German) for the ConversionViewController labels & text, which were created via the storyboard, and then editing Localizable.strings (German) for the MapViewController button titles, which were created programatically. I omitted the workaround for the Xcode bug that Christian described in the book, reasoning that Xcode may no longer have that issue. I just set the scheme for German and Germany, and then built the app, and ran it, and it works - all of my labels and button titles are in German now!